Fix binary reading of tuples containing non-nullable types (#3734)
We must write them to a tuple with nullable types, then fix that up when
reading. This is similar to what we do in handleNonNullableLocals, except
that it operates on the entire tuple type, so it can't share that code.

This fixes a regression from #3710 that was harder to notice by the fuzzer
until now.