Run spec test all at once after binary transform (#3817)
#3792 added support for module linking and (register command to
wasm-shell, but forgot about three problems:

-    Splitting spec tests prevents linking test modules together.
-    Registered modules may still be used in assertions or an invoke
-    Modules may re-export imported objects

This PR appends transformed modules after binary checks to a spec.wast
file, plus assertion tests and register commands. Then runs wasm-shell
on the whole file. It also keeps both the module name and its registered
name available in wasm-shell for use in shell commands and linked
modules. Furthermore, it correctly finds the module where an object is
defined even if it is imported and re-exported several times.

The updated version of imports.wast spec test is enabled to verify the
fixes.