Refactor Child*Iterator for simplicity and to allow modifications. NFC (#3826)
Instead of creating a class and doing a traversal, simply use
wasm-delegates-fields to get the children directly. This is simpler and
faster.

This requires a new way to override the behavior by ChildValueIterator.
I ended up using CRTP.

This stores pointers to the children instead of the children. This will
allow modifying the children using these classes (which is the reason
I started this refactoring - next PR will use it). This is also slightly
faster in theory as we avoid doing loads from memory to find the
children (which maybe we never get to doing if the iteration is stopped
early).

Not the goal here, but this speeds up roundtripping by 12%.

This is NFC as nothing uses the new ability to modify things yet.