Generalize moving of identical code from if/select arms (#3833)
Effects are fine in the moved code, if we are doing so on an if
(which runs just one arm anyhow).

Allow unreachable, which lets us hoist returns for example.

Allow none, which lets us hoist drop and call for example. For
this we also need to be careful with subtyping, as at least drop
is polymorphic, so the child types may not have an LUB (see
example in code).

Adds a small ShallowEffectAnalyzer child of EffectAnalyzer that
calls visit to just do a shallow analysis (instead of walk which
walks the children).