OptimizeInstructions: Fix/ignore eqz hoisting of if with unreachable arm (#3835)
We tried to ignore unreachable code, but only checked the type of
the entire node. But an arm might be unreachable, and after moving
code around that requires more work to update the type. But such
cases are best left to DCE anyhow, so just check for any unreachability
and stop there.