Fix a new clang warning on implicit Tuple construction (#3841)
Fixes #3838 and current breakage on CI here.

IIUC the warning is that we define a copy assignment operator, but
did not define a copy constructor. Clang wants both to exist now. However,
we delete the copy assignment one, so perhaps we should find a way to
avoid a copy on construction too? That happens here:

../src/wasm/wasm-type.cpp:60:51: note: in implicit copy constructor for 'wasm::Tuple' first required here
  TypeInfo(const Tuple& tuple) : kind(TupleKind), tuple(tuple) {}
                                                  ^
But I don't see an easy way to remove the copy there, so this PR
defines the copy constructor as clang wants.