Add pass argument sugar to commandline (#3882)
We have --pass-arg that allows sending an argument to a pass, like this:

wasm-opt --do-stuff --pass-arg=do-stuff@FUNCTION_NAME

With this PR that is equivalent to this:

wasm-opt --do-stuff=FUNCTION_NAME

That is,one can just give an argument to a pass on the commandline.

This fixes the Optional mode in command-line.h/cpp. That was not
actually used anywhere before this PR.

Also rename --extract-function's pass argument to match it. That is, the usage
used to be

wasm-opt --extract-function --pass-arg=extract@FUNCTION_NAME

Note how the pass name differed from the pass-arg name. This changes it to
match. This is a breaking change, but I doubt this is used enough to justify
any deprecation / backwards compatibility effort, and any usage is almost
certainly manual, and with PR writing it manually becomes easier as one
can do

wasm-opt --extract-function=FUNCTION_NAME

The existing test for that is kept (&renamed), and a new test added to test the
new notation.

This is a step towards unifying the symbol map functionality between
wasm-as and wasm-opt (later PRs will turn the symbol mapping pass into
a pass that receives an argument).