Support --symbolmap and --symbolmap=FOO in wasm-opt (#3885)
wasm-as supports --symbolmap=FOO as an argument. We got a request to
support the same in wasm-opt. wasm-opt does have --print-function-map which
does the same, but as a pass. To unify them, use the new pass arg sugar from
#3882 which allows us to add a --symbolmap pass whose argument can be
set as --symbolmap=FOO. That perfectly matches the wasm-as notation.

For now, keep the old --print-function-map notation as well, to not break
emscripten. After we remove it there we can remove it here.