Handle invokes of invalid function pointers. See #14174 (#3951)
PostEmscripten will turn an invoke of a constant function
pointer index into a direct call. However, due to UB it is possible to
have invalid function pointers, and we should not crash on that
(and do nothing to optimize, of course).

Mostly whitespace; to avoid deep nesting, I added more
early returns.